requirements.txt
# Core Python Requirements for NOC Buzzer System Dashboard

# Web Framework and Dashboard
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7

# Database Connectivity
mysql-connector-python==8.1.0

# Gmail Integration
simplegmail==4.1.1
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0

# Modbus Communication
pyModbusTCP==0.2.0

# Machine Learning and Data Processing
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
scipy==1.11.1

# Natural Language Processing
nltk==3.8.1

# Email Processing
imap-tools==1.5.0
email-validator==2.0.0

# HTML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP Requests
requests==2.31.0
urllib3==2.0.4

# Text Processing and Similarity
difflib2==0.1.3

# Data Serialization
jsonschema==4.19.0

# Logging and Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing (Optional)
pytest==7.4.2
pytest-flask==1.2.0

# Additional Dependencies from your existing code
# OpenAI API (if using ChatgptAPI.py)
openai==0.27.8

# Email libraries (for different email reading methods)
imaplib2==3.6

# Time and scheduling
schedule==1.2.0

# Configuration management
python-dotenv==1.0.0

# Async support (if needed)
asyncio==3.4.3

# Security
cryptography==41.0.4

# Threading and multiprocessing support
threading-timer==0.1.0

# Essential requirements for NOC Buzzer Dashboard
flask==2.3.3
flask-cors==4.0.1
mysql-connector-python==8.2.0
requests==2.31.0
beautifulsoup4==4.12.2

# Your existing dependencies (likely already installed)
simplegmail>=4.0.0
pyModbusTCP>=0.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
nltk>=3.8.0